requirements.txt
# Core Data Processing
pyspark>=3.4.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Google Cloud
google-cloud-bigquery>=3.11.0
google-cloud-storage>=2.10.0
google-cloud-pubsub>=2.30.0

# Dashboard
streamlit>=1.32.0
plotly>=5.19.0
db-dtypes>=1.2.0

# Testing
pytest>=7.4.0
pytest-spark>=0.6.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.14.0
